perf(rust, python): [csv] SIMD accelerate SplitFields `-40%`
