+++ polars/polars-io/src/lib.rs
@@ -1,4 +1,5 @@
 #![cfg_attr(docsrs, feature(doc_auto_cfg))]
+#![cfg_attr(feature = "simd", feature(portable_simd))]
 
 #[cfg(feature = "avro")]
 pub mod avro;
